Introduce a homepage for card based frameworks Replacing the default "click a framework" button with a markdown-based home page for each framework. This is optional.
