+++ frameworks/design/user-research.md
@@ -5,6 +5,7 @@ sidebarTitle: "🎯 User Research"
 sidebarGroup: "design"
 yaml: false
 levels:
+homepage:
 ---
 ## Level 1
