+++ frameworks/design/visual-design.md
@@ -5,6 +5,7 @@ sidebarTitle: "🎯 Visual Design"
 sidebarGroup: "design"
 yaml: false
 levels:
+homepage:
 ---
 ## Level 1
