+++ frameworks/engineering/data.md
@@ -5,6 +5,7 @@ sidebarTitle: "📊 Data"
 sidebarGroup: "engineering"
 yaml: true
 levels: 6
+homepage: false
 topics:
   - name: "communication"
     title:
