+++ frameworks/engineering/mobile.md
@@ -5,6 +5,7 @@ sidebarTitle: "👩‍💻 Mobile"
 sidebarGroup: "engineering"
 yaml: true
 levels: 6
+homepage: false
 topics:
   - name: "communication"
     title:
