+++ frameworks/generic.md
@@ -5,6 +5,7 @@ sidebarTitle: "📄 Generic"
 sidebarGroup:
 yaml: true
 levels: 6
+homepage: false
 topics:
   - name: "knowledge"
     title: "🧠 Knowledge"
