+++ src/views/FrameworkView.js
@@ -26,6 +26,7 @@ class View extends React.Component<Props> {
           <LevelledRenderer
             pageData={pageFrontmatter}
             genericData={genericFrontmatter}
+            html={html}
           />
         ) : (
           <TextRenderer data={pageFrontmatter} html={html} />
@@ -63,6 +64,7 @@ export const pageQuery = graphql`
         sidebarGroup
         yaml
         levels
+        homepage
         topics @include(if: $isYaml) {
           name
           title
@@ -87,6 +89,7 @@ export const pageQuery = graphql`
         sidebarGroup
         yaml
         levels
+        homepage
         topics @include(if: $isYaml) {
           name
           title
